Added Moving average to the gp IR
