--- src/org/usfirst/frc/team3316/robot/stacker/GamePieceType.java
@@ -1,7 +1,0 @@
-
-package org.usfirst.frc.team3316.robot.stacker;
-
-public enum GamePieceType 
-{
-	Container, Tote;
-}
