+++ src/org/usfirst/frc/team3316/robot/utils/GamePieceCollected.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team3316.robot.rollerGripper;
+package org.usfirst.frc.team3316.robot.utils;
 
 public enum GamePieceCollected 
 {
